Rename location-tools to axis-ui
