+++ src-js/views-fontoverview/src/panel-navigation.js
@@ -1,3 +1,7 @@
+import {
+  makeFontAxisAccordionItems,
+  setShowEffectiveLocationDefaults,
+} from "@fontra/core/axis-ui.js";
 import { groupByKeys, groupByProperties } from "@fontra/core/glyph-organizer.js";
 import {
   CheckboxGroup,
@@ -6,10 +10,6 @@ import {
 } from "@fontra/core/glyphsets-ui.js";
 import * as html from "@fontra/core/html-utils.js";
 import { translate } from "@fontra/core/localization.js";
-import {
-  makeFontAxisAccordionItems,
-  setShowEffectiveLocationDefaults,
-} from "@fontra/core/location-tools.js";
 import { ObservableController } from "@fontra/core/observable-object.ts";
 import { difference, symmetricDifference, union } from "@fontra/core/set-ops.js";
 import { popupSelect } from "@fontra/core/ui-utils.js";
